+++ lib/cuke_modeler/adapters/gherkin_18_adapter.rb
@@ -9,5 +9,15 @@ module CukeModeler
 
   class Gherkin18Adapter < Gherkin9Adapter
 
+    # Adapts the AST sub-tree that is rooted at the given rule node.
+    def adapt_rule(rule_ast)
+      adapted_rule = super(rule_ast)
+
+      # Tagging of Rules was added in Gherkin 18
+      adapted_rule['tags'] = adapt_tags(rule_ast[:rule])
+
+      adapted_rule
+    end
+
   end
 end