+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-Noting yet...
+### Added
+ - `Rule` models are now taggable elements
+
 
 ## [3.8.0] - 2021-04-18